task_1/requirements.txt
pyyaml
pillow
torch
torchvision
python-box
ipykernel
ipywidgets
notebook
jupyterlab
matplotlib
seaborn
numpy
tqdm
tensorboard
rich
drf-spectacular
drf-spectacular[sidecar]
opencv-python
torchinfo
scikit-learn
timm